przenioslem movie kontroler do film controler i zmienilem nazwe z mobierepository na filmrepository
